requirements.txt
# Shared backend dependencies
Flask==2.3.3
Flask-Cors==4.0.1
python-dotenv==1.0.0
argon2-cffi==23.1.0
PyJWT==2.8.0
requests==2.31.0